Upon further investigation, DWORD *needs* to be defined as a 32 bit unsigned integer; several codec functions use DWORD and WORD to assume 32 bit and 16 values respectively. So DWORD is uint32_t. This causes a problem in the messaging code where pointers are sent around using the DWORD dwParamA, and dwParamB. This is *so* bad; but it's there now and will be hard to route out. For now, I changed the type of all dwParamA, dwParamB to my own custom Param_t type. This is a DWORD on Windows and size_t on macOS. This should keep pointers from being truncated as they are passed around. Yikes.

+++ nTA/Source/std_Platform.h
@@ -56,7 +56,7 @@ typedef void * LPVOID;
 typedef const char * LPCTSTR;
 typedef char * LPTSTR;
 typedef char * LPSTR;
-typedef size_t DWORD;
+typedef uint32_t DWORD;
 typedef uint16_t WORD;
 typedef uint8_t BYTE;
 typedef uint32_t UINT32;
@@ -65,7 +65,7 @@ typedef uint16_t UINT16;
 typedef int16_t INT16;
 typedef uint8_t UINT8;
 typedef int8_t INT8;
-typedef unsigned int UINT;
+typedef uint32_t UINT;
 typedef long LRESULT;
 #define stricmp strcasecmp
 #define strnicmp strncasecmp
